requirements.txt
gputil==1.4.0
hydra-core==1.3.2
hydra-joblib-launcher
numpy==2.3.1
omegaconf==2.3.0
openai==1.93.1
pandas==2.3.0
psutil==7.0.0
sentence_transformers==4.1.0
tqdm==4.67.1
gritlm
xlsxwriter
xgboost
openml
hdbscan